Re-structure java samples
