--- dop-model-data-as-data/java/model-data-as-data/src/main/java/dop/example/Main.java
@@ -1,18 +1,0 @@
-package dop.example;
-
-enum UserType {
-    ADMIN, CUSTOMER, GUEST
-}
-record User(int id, String name, UserType userType) {
-    public User(int id, String name) {
-        this(id, name, UserType.GUEST);
-    }
-}
-
-public class Main {
-    public static void main(String[] args) {
-        User customer = new User(1, "John Doe");
-        System.out.printf("User '%s' with id '%s' as '%s' created successfully",
-                customer.name(), customer.id(), customer.userType());
-    }
-}
